backend/requirements.txt
fastapi
uvicorn
supabase
google-generativeai
stripe
python-dotenv
pymupdf